Consolidate type.c and utype.c
